requirements.txt
PySensors
pyserial
psutil

